requirements.txt
# Image Analyzer and Summarizer dependencies

# Core dependencies
ollama>=0.1.26
pillow>=9.0.0      # For image handling, validation, and chunking
requests>=2.28.0   # For API requests
